Add child pages, points adjustment, and settings page. De-lint python code.

+++ chornado_app/parents.py
@@ -1,16 +1,21 @@
 from flask import redirect, render_template, request, Blueprint, flash
-from flask_login import login_required
+from flask_login import login_required, current_user
+from werkzeug.security import check_password_hash, generate_password_hash
 
-from .forms import *
-from .sql_models import *
-from .helpers import *
+from .forms import (AssignedChoreForm, RewardForm, RegisterForm, PointsForm,
+    ChoreForm, ParentResetPasswordForm)
+from .sql_models import (db, User, AssignedChore, Chore, Reward, Notification)
+from .helpers import (db_commit, redirect_url, register_user, flash_errors,
+    approve_completed, reject_completed, create_chore, assign_chore, edit_chore,
+    create_reward, edit_reward)
 
-parents_bp = Blueprint('parents', __name__, url_prefix="/parent")
+parent_bp = Blueprint('parent', __name__, url_prefix="/parent")
 
-@parents_bp.route('/home')
+@parent_bp.route('/home')
 @login_required
 def home():
     """Parents homepage"""
+
     parent = current_user
     # Get list of children belonging to user
     children = User.query.filter_by(parent=parent.id).order_by(User.first_name)        
@@ -19,7 +24,7 @@ def home():
     notifications = parent.notifications
     chore_form = AssignedChoreForm()
     reward_form = RewardForm()
-    
+
     # Create list of dictionaries of data for children
     # and count of the chores assigned to them
     for child in children:
@@ -29,16 +34,17 @@ def home():
         child_data.append(child_dict)
     return render_template('parents/home.html', child_data=child_data,
     notifications=notifications, chore_form=chore_form, reward_form=reward_form)
-    
-@parents_bp.route('/children', methods=['GET', 'POST'])
+
+@parent_bp.route('/children', methods=['GET', 'POST'])
 @login_required
 def children():
     """Parent's children page for listing and creating child accounts"""
-    
+
     parent = current_user
     register_form = RegisterForm()
+    points_form = PointsForm()
     chore_form = AssignedChoreForm()
-    
+
     if request.method == 'POST':
         # Register new child account
         if register_form.validate() and register_form.submit.data:
@@ -52,109 +58,121 @@ def children():
             db_commit()
             return redirect(redirect_url())
 
+        # Adjust child's points
+        if points_form.validate() and points_form.adjust.data:
+            child = User.query.get(points_form.child_id.data)
+            new_points = points_form.points.data
+            child.points += new_points
+            if new_points < 0:
+                new_points = -new_points
+                flash(f"{new_points} points removed from {child.first_name}'s account.",
+                    'success')
+            else:
+                flash(f"{new_points} points added to {child.first_name}'s account.",
+                    'success')
+            db_commit()
+            return redirect(redirect_url())
+
         # Check if chore_form is being submitted
-        elif (chore_form.validate() and (chore_form.delete.data
+        if (chore_form.validate() and (chore_form.delete.data
             or chore_form.approve.data or chore_form.reject.data)):
-            
+
             current_chore = AssignedChore.query.get(chore_form.chore_id.data)
             chore_info = Chore.query.get(current_chore.chore_id)
             child = User.query.get(current_chore.user_id)
-        
+
             # Approve completed chore, assign points, and delete assigned chore
             if chore_form.approve.data:
                 new_points = chore_info.value
                 approve_completed(current_chore, child, new_points)
                 flash(f'{chore_info.name} complete!', 'success')
-            
+
             # Delete assigned chore
             elif chore_form.delete.data:
                 db.session.delete(current_chore)
                 flash(f'{chore_info.name} has been removed for {child.first_name}', 'success')
-            
+
             # Reject completed chore, and set status back to active
             elif chore_form.reject.data:
                 reject_completed(current_chore)
                 flash(f'{chore_info.name} has been sent back to {child.first_name}', 'success')
             db_commit()
             return redirect(redirect_url())
-        
-        else:
-            flash_errors(register_form)
-            flash_errors(chore_form)
-            return redirect(redirect_url())
-    
-    else:
-        children = User.query.filter_by(parent=parent.id).order_by(User.first_name)
-        child_data = []
 
-        # Create list of dictionaries of data for children
-        # and the chores assigned to them
-        for child in children:
-            chores = []
-            assigned_chores = child.assigned_chores.order_by(AssignedChore.id)
-            for assigned_chore in assigned_chores:
-                chore_info = Chore.query.get(assigned_chore.chore_id)
-                chore_dict = {'id': assigned_chore.id, 'name': chore_info.name,
-                    'points': chore_info.value, 'state': assigned_chore.state}
-                
-                chores.append(chore_dict)
-            
-            child_dict = {'username': child.username, 'first_name': child.first_name,
-                'points': child.points, 'id': child.id, 'chores': chores}
+        flash_errors(register_form)
+        flash_errors(chore_form)
+        return redirect(redirect_url())
 
-            child_data.append(child_dict)
-        
-        return render_template('parents/children.html', template_form=register_form,
-        chore_form=chore_form, child_data=child_data)
+    children = User.query.filter_by(parent=parent.id).order_by(User.first_name)
+    child_data = []
 
-@parents_bp.route('/chores', methods=['GET', 'POST'])
+    # Create list of dictionaries of data for children
+    # and the chores assigned to them
+    for child in children:
+        chores = []
+        assigned_chores = child.assigned_chores.order_by(AssignedChore.id)
+        for assigned_chore in assigned_chores:
+            chore_info = Chore.query.get(assigned_chore.chore_id)
+            chore_dict = {'id': assigned_chore.id, 'name': chore_info.name,
+                'points': chore_info.value, 'state': assigned_chore.state}
+
+            chores.append(chore_dict)
+
+        child_dict = {'username': child.username, 'first_name': child.first_name,
+            'points': child.points, 'id': child.id, 'chores': chores, 'count': len(chores)}
+
+        child_data.append(child_dict)
+
+    return render_template('parents/children.html', template_form=register_form,
+    points_form=points_form, chore_form=chore_form, child_data=child_data)
+
+@parent_bp.route('/chores', methods=['GET', 'POST'])
 @login_required
 def parent_chores():
     """Parent's chore page for creating, viewing, editing, and assigning chores"""
-    
+
     parent = current_user
     form = ChoreForm()
-    
+
     if request.method == 'POST' and form.validate_on_submit():
         name = form.name.data
         value = form.value.data
         child = User.query.get(form.child.data)
         chore_id = form.chore_id.data
         chore = Chore.query.get(chore_id)
-        
+
         # Create new chore
         if form.create.data:
             create_chore(name, value, parent.id)
-        
+
         # Assign chore to child
         elif form.assign.data:
             assign_chore(chore_id, child.id)
             flash(f'{chore.name} assigned to {child.first_name}', 'success')
-        
+
         # Edit existing chore
         elif form.edit.data:
-            edit_chore(chore, name, value)            
+            edit_chore(chore, name, value)
             flash(f'{chore.name} has been updated', 'success')
-        
-        # Delete chore        
+
+        # Delete chore
         elif form.delete.data:
             db.session.delete(chore)
         db_commit()
         return redirect(redirect_url())
-    
-    else:
-        flash_errors(form)
-        chores = parent.chores.order_by(Chore.name)
-        return render_template('parents/chores.html', template_form=form, chores=chores)
 
-@parents_bp.route('/rewards', methods=['GET', 'POST'])
+    flash_errors(form)
+    chores = parent.chores.order_by(Chore.name)
+    return render_template('parents/chores.html', template_form=form, chores=chores)
+
+@parent_bp.route('/rewards', methods=['GET', 'POST'])
 @login_required
 def parent_rewards():
     """Route for viewing, creating, and editing rewards"""
 
     parent = current_user
-    form = RewardForm()    
-        
+    form = RewardForm()
+
     if request.method == 'POST' and form.validate_on_submit():
         name = form.name.data
         cost = form.cost.data
@@ -168,11 +186,41 @@ def parent_rewards():
         elif form.deliver.data:
             notification = Notification.query.get(form.notification_id.data)
             db.session.delete(notification)
-            flash(f'Reward delivered', 'success')
+            message = f'You have been given {reward.name}!'
+            new_notification = Notification(type='reward', message=message,
+                user_id=notification.child_id, child_id=notification.child_id,
+                reward_id=notification.reward_id)
+            db.session.add(new_notification)
+            flash('Reward delivered', 'success')
         db_commit()
         return redirect(redirect_url())
-    else:
-        flash_errors(form)
-        rewards = parent.rewards.order_by(Reward.name)
-        return render_template('parents/rewards.html', template_form=form,
-        rewards=rewards)
+
+    flash_errors(form)
+    rewards = parent.rewards.order_by(Reward.name)
+    return render_template('parents/rewards.html', template_form=form,
+    rewards=rewards)
+
+@parent_bp.route('/settings', methods=['POST', 'GET'])
+@login_required
+def settings():
+    """Parent's settings page"""
+
+    parent = current_user
+    form = ParentResetPasswordForm()
+
+    if request.method == 'POST' and form.validate_on_submit():
+        old_password = form.old_password.data
+        if check_password_hash(parent.password_hash, old_password):
+            new_password = form.new_password.data
+            new_hash = generate_password_hash(new_password,
+                method='pbkdf2:sha256', salt_length=8)
+
+            parent.password_hash = new_hash
+            db_commit()
+            flash('Password changed','success')
+        else:
+            flash('Please enter correct old password.', 'error')
+        return redirect(redirect_url())
+
+    flash_errors(form)
+    return render_template('parents/settings.html', parent=parent, template_form=form)
